+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Version 2.0.4 *(2026-[M]-[D])*
 -----------------------------
 
 * Implement `del` statement with identifiers.
+* Implement `try ... except ... else ... finally` statement.
 
 Version 2.0.3 *(2026-06-16)*
 -----------------------------